Created By: @李恒
Reviewed By: @李元明 @廖凯
Approved By: @廖凯 
Accepted By: @李恒
URL: https://devops.codingcorp.net/p/platform/d/platform-project/git/merge/14586

+++ lib/src/main/java/net/coding/lib/project/dto/request/ProjectMemberQueryPageReqDTO.java
@@ -20,7 +20,7 @@ public class ProjectMemberQueryPageReqDTO {
     @ApiModelProperty(value = "项目Id")
     private Integer projectId;
     @ApiModelProperty(value = "权限组Id")
-    private Integer policyId;
+    private Long policyId;
     @ApiModelProperty(value = "关键字搜索")
     private String keyword;
 }
